+++ scylla/src/session/session_builder.rs
@@ -267,7 +267,7 @@ impl GenericSessionBuilder<DefaultMode> {
     /// # use std::sync::Arc;
     /// # use scylla::{Session, SessionBuilder};
     /// # use scylla::connection::{AddressTranslator, TranslationError};
-    /// # use scylla::transport::metadata::UntranslatedPeer;
+    /// # use scylla::cluster::metadata::UntranslatedPeer;
     /// struct IdentityTranslator;
     ///
     /// #[async_trait]
