+++ scylla/src/session/session_test.rs
@@ -1,4 +1,6 @@
 use crate::batch::{Batch, BatchStatement};
+use crate::cluster::metadata::Strategy::NetworkTopologyStrategy;
+use crate::cluster::metadata::{CollectionType, ColumnKind, CqlType, NativeType, UserDefinedType};
 use crate::deserialize::DeserializeOwnedValue;
 use crate::prepared_statement::PreparedStatement;
 use crate::query::Query;
@@ -8,10 +10,6 @@ use crate::session::Session;
 use crate::statement::Consistency;
 use crate::tracing::TracingInfo;
 use crate::transport::errors::{BadKeyspaceName, BadQuery, DbError, QueryError};
-use crate::transport::metadata::Strategy::NetworkTopologyStrategy;
-use crate::transport::metadata::{
-    CollectionType, ColumnKind, CqlType, NativeType, UserDefinedType,
-};
 use crate::transport::partitioner::{
     calculate_token_for_partition_key, Murmur3Partitioner, Partitioner, PartitionerName,
 };
