+++ scylla/src/transport/cluster.rs
@@ -6,11 +6,8 @@ use crate::routing::{Shard, Token};
 use crate::session::TABLET_CHANNEL_SIZE;
 use crate::transport::errors::{BadQuery, NewSessionError, QueryError};
 use crate::transport::host_filter::HostFilter;
-use crate::transport::{
-    metadata::{Keyspace, Metadata, MetadataReader},
-    node::Node,
-    partitioner::PartitionerName,
-};
+use crate::transport::node::Node;
+use crate::transport::partitioner::PartitionerName;
 
 use arc_swap::ArcSwap;
 use futures::future::join_all;
@@ -27,10 +24,11 @@ use uuid::Uuid;
 
 use super::locator::tablets::{RawTablet, Tablet, TabletsInfo};
 use super::node::{InternalKnownNode, NodeAddr};
+
 use super::NodeRef;
+use crate::cluster::metadata::{Keyspace, Metadata, MetadataReader, Strategy};
 
 use super::locator::ReplicaLocator;
-use super::metadata::Strategy;
 use super::partitioner::calculate_token_for_partition_key;
 
 /// Cluster manages up to date information and connections to database nodes.
